Split diff runtime modules
